REF: move Block.astype implementation to array_algos
